greatly improve the delet page UI

+++ acp/pages_module.php
@@ -63,8 +63,21 @@ class pages_module
 			break;
 
 			case 'delete':
-				// Delete a page
-				$admin_controller->delete_page($page_id);
+				// Use a confirm box routine when deleting a page
+				if (confirm_box(true))
+				{
+					// Delete page on confirmation from the user
+					$admin_controller->delete_page($page_id);
+				}
+				else
+				{
+					// Request confirmation from the user to delete the page
+					confirm_box(false, $user->lang('ACP_PAGE_DELETE_CONFIRM'), build_hidden_fields(array(
+						'page_id'	=> $page_id,
+						'mode'		=> $mode,
+						'action'	=> $action))
+					);
+				}
 			break;
 
 			case 'purge_icons':
